requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudevents==1.11.0
comm==0.2.3
db-dtypes==1.4.3
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dlt==1.15.0
executing==2.2.0
flask==3.1.2
frozenlist==1.7.0
fsspec==2025.7.0
functions-framework==3.9.2
gcloud==0.18.3
gcsfs==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
giturlparse==0.12.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.36.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.13.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.102.0
orjson==3.11.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathvalidate==3.3.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requirements-parser==0.13.0
rich==14.1.0
rich-argparse==1.7.1
rpds-py==0.27.1
rsa==4.9.1
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlglot==27.8.0
stack-data==0.6.3
starlette==0.47.3
streamlit==1.49.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvicorn-worker==0.3.0
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.1.3
yarl==1.20.1
